requirements.txt
Django~=3.1.3
djangorestframework-bulk~=0.2.1
djangorestframework~=3.12.2
django-filter~=2.4.0
django-model-utils~=4.0.0
pytest~=6.1.2